requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
fastparquet==2023.10.1
garth==0.4.44
google-cloud-logging==3.9.0
lxml==5.1.0
numpy==1.26.3
pandas==2.2.0
pyarrow==15.0.0
python-dotenv==1.0.0
